+++ docs/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 6.5.0 Features
 
+- `[Searchfield]` Exposed the `collapseSize` setting for use with the IDS Enteprise Searchfield component ([#719](https://github.com/infor-design/enterprise-ng/issues/719))
+
 ## v6.4.0
 
 ### 6.4.0 Breaking Changes